+++ main.go
@@ -14,7 +14,6 @@ import (
 
 	"github.com/jackc/pgx/v4/pgxpool"
 	"github.com/pkg/errors"
-	"github.com/streadway/amqp"
 )
 
 func main() {
@@ -30,7 +29,8 @@ func main() {
 	}
 
 	configFlag := flag.String("config", "", "Select config .json file.")
-	serviceFlag := flag.String("service", "", fmt.Sprintf("Select one service for activate.\nAvailable: %s.", strings.Join(services, ", ")))
+	serviceFlag := flag.String("service", "", fmt.Sprintf("Select single service for activate.\nAvailable: %s.", strings.Join(services, ", ")))
+	workersFlag := flag.Int("workers", 1, "Select number of running workers of the service")
 	flag.Parse()
 	if *serviceFlag == "" || !isExist(*serviceFlag, services) {
 		_, _ = fmt.Fprintf(os.Stderr, "Select one service for activate (flag -service)\n")
@@ -58,22 +58,6 @@ func main() {
 		os.Exit(1)
 	}
 
-	rabbitConn, err := amqp.Dial(conf.RabbitURL)
-	if err != nil {
-		loki.errorWithExit(errors.WithMessage(errors.WithStack(err), "Can't connect to rabbitMQ"))
-	}
-	defer func(conn *amqp.Connection) {
-		err := conn.Close()
-		if err != nil {
-			loki.errorWithExit(errors.WithMessage(errors.WithStack(err), "Can't close connection to rabbitMQ"))
-		}
-	}(rabbitConn)
-
-	queues := []string{queueSearchQueries, queueSearchBodiesFromDDG, queueProxySources, queueProxySourceHTML, queueRawProxies, queueCheckProxies}
-	if err := initQueues(rabbitConn, queues); err != nil {
-		loki.errorWithExit(errors.WithMessage(errors.WithStack(err), "Can't init queues"))
-	}
-
 	rdbForSites, err := newRedisDB(conf.RedisURLForSites)
 	if err != nil {
 		loki.errorWithExit(errors.WithMessage(errors.WithStack(err), "Can't connect to redis for site urls"))
@@ -91,7 +75,6 @@ func main() {
 
 	app := App{
 		loki:          loki,
-		rabbitConn:    rabbitConn,
 		postgresPool:  postgresPool,
 		rdbForSites:   rdbForSites,
 		rdbForProxies: rdbForProxies,
@@ -101,85 +84,75 @@ func main() {
 	exit := make(chan os.Signal, 1)
 	signal.Notify(exit, syscall.SIGINT, syscall.SIGTERM)
 
-	if *serviceFlag == serviceFillSearchQueries {
-		loki.info("worker started")
-		go func() {
+	for i := 0; i < *workersFlag; i++ {
+		loki.info(fmt.Sprintf("Service started (worker %d)", i))
+
+		if *serviceFlag == serviceFillSearchQueries {
 			fromProxies := false
-			for {
-				if err := fillSearchQueries(app, fromProxies); err != nil {
-					loki.errorWithExit(errors.WithStack(err))
+			queueDest := NewRabbitMQSession(app.loki, queueSearchQueries, app.conf.RabbitURL)
+
+			go func() {
+				for {
+					if err := fillSearchQueries(app, queueDest, fromProxies); err != nil {
+						loki.error(errors.WithStack(err))
+					}
+					fromProxies = !fromProxies
+					time.Sleep(time.Minute)
 				}
-				fromProxies = !fromProxies
-				time.Sleep(time.Minute)
-			}
-		}()
-	}
+			}()
+		}
 
-	if *serviceFlag == serviceSendSearchBodyFromDDGToQueue {
-		loki.info("worker started")
-		go func() {
-			if err := sendSearchBodyFromDDGToQueue(app, proxyURL, userAgent); err != nil {
-				loki.errorWithExit(errors.WithStack(err))
-			}
-		}()
-	}
+		if *serviceFlag == serviceSendSearchBodyFromDDGToQueue {
+			queueSource := NewRabbitMQSession(app.loki, queueSearchQueries, app.conf.RabbitURL)
+			queueDest := NewRabbitMQSession(app.loki, queueSearchBodiesFromDDG, app.conf.RabbitURL)
+			go sendSearchBodyFromDDGToQueue(app, queueSource, queueDest, proxyURL, userAgent)
+		}
 
-	if *serviceFlag == serviceProcessSearchBodyFromDDG {
-		loki.info("worker started")
-		go func() {
-			if err := processSearchBodyFromDDG(app); err != nil {
-				loki.errorWithExit(errors.WithStack(err))
-			}
-		}()
-	}
+		if *serviceFlag == serviceProcessSearchBodyFromDDG {
+			queueSource := NewRabbitMQSession(app.loki, queueSearchBodiesFromDDG, app.conf.RabbitURL)
+			queueDest := NewRabbitMQSession(app.loki, queueProxySources, app.conf.RabbitURL)
+			go processSearchBodyFromDDG(app, queueSource, queueDest)
+		}
 
-	if *serviceFlag == serviceSendHTMLFromProxySourceToQueue {
-		loki.info("worker started")
-		go func() {
-			if err := sendHTMLFromProxySourceToQueue(app); err != nil {
-				loki.errorWithExit(errors.WithStack(err))
-			}
-		}()
-	}
+		if *serviceFlag == serviceSendHTMLFromProxySourceToQueue {
+			queueSource := NewRabbitMQSession(app.loki, queueProxySources, app.conf.RabbitURL)
+			queueDest := NewRabbitMQSession(app.loki, queueProxySourceHTML, app.conf.RabbitURL)
+			go sendHTMLFromProxySourceToQueue(app, queueSource, queueDest)
+		}
 
-	if *serviceFlag == serviceProcessSourceHTML {
-		loki.info("worker started")
-		go func() {
-			if err := processSourceHTML(app); err != nil {
-				loki.errorWithExit(errors.WithStack(err))
-			}
-		}()
-	}
+		if *serviceFlag == serviceProcessSourceHTML {
+			queueSource := NewRabbitMQSession(app.loki, queueProxySourceHTML, app.conf.RabbitURL)
+			chDestRawProxies := NewRabbitMQSession(app.loki, queueRawProxies, app.conf.RabbitURL)
+			chDestProxySources := NewRabbitMQSession(app.loki, queueProxySources, app.conf.RabbitURL)
+			go processSourceHTML(app, queueSource, chDestRawProxies, chDestProxySources)
+		}
 
-	if *serviceFlag == serviceProcessRawProxy {
-		loki.info("worker started")
-		go func() {
-			if err := processRawProxy(app); err != nil {
-				loki.errorWithExit(errors.WithStack(err))
-			}
-		}()
-	}
+		if *serviceFlag == serviceProcessRawProxy {
+			queueSource := NewRabbitMQSession(app.loki, queueRawProxies, app.conf.RabbitURL)
+			go processRawProxy(app, queueSource)
+		}
 
-	if *serviceFlag == serviceFillCheckProxiesQueue {
-		loki.info("worker started")
-		go func() {
-			for {
-				if err := fillCheckProxiesQueue(app); err != nil {
-					loki.errorWithExit(errors.WithStack(err))
+		if *serviceFlag == serviceFillCheckProxiesQueue {
+			queueDest := NewRabbitMQSession(app.loki, queueCheckProxies, app.conf.RabbitURL)
+			go func() {
+				for {
+					if err := fillCheckProxiesQueue(app, queueDest); err != nil {
+						loki.error(errors.WithStack(err))
+					}
+					time.Sleep(10 * time.Minute)
 				}
-				time.Sleep(time.Minute * 10)
-			}
-		}()
+			}()
+		}
+
+		if *serviceFlag == serviceProcessCheckProxies {
+			queueSource := NewRabbitMQSession(app.loki, queueCheckProxies, app.conf.RabbitURL)
+			go processCheckProxy(app, queueSource)
+		}
 	}
 
-	if *serviceFlag == serviceProcessCheckProxies {
-		loki.info("worker started")
-		go func() {
-			if err := processCheckProxy(app); err != nil {
-				loki.errorWithExit(errors.WithStack(err))
-			}
-		}()
+	if shutdown := <-exit; shutdown != nil {
+		app.loki.info("Service shutdown")
+		time.Sleep(lokiBatchWait + time.Second*2)
+		os.Exit(1)
 	}
-
-	<-exit
 }