+++ service.go
@@ -9,21 +9,10 @@ import (
 	"time"
 
 	"github.com/pkg/errors"
-	"github.com/streadway/amqp"
 )
 
-func fillSearchQueries(app App, fromProxies bool) error {
+func fillSearchQueries(app App, queueDest *RabbitMQSession, fromProxies bool) error {
 	startTime := time.Now()
-
-	ch, err := app.rabbitConn.Channel()
-	if err != nil {
-		return err
-	}
-	defer func(ch *amqp.Channel) {
-		if err := ch.Close(); err != nil {
-			app.loki.error(errors.WithStack(err))
-		}
-	}(ch)
 
 	var q string
 	proxies := make([]string, 0)
@@ -34,7 +23,7 @@ func fillSearchQueries(app App, fromProxies bool) error {
 				proxies = append(proxies, fProxy.URL)
 			}
 		} else {
-			app.loki.error(errors.WithStack(err))
+			return err
 		}
 	}
 
@@ -45,36 +34,16 @@ func fillSearchQueries(app App, fromProxies bool) error {
 		q = randomElementFromArray(queries)
 	}
 
-	if err := publish(ch, queueSearchQueries, []byte(q)); err != nil {
+	if err := queueDest.Push([]byte(q)); err != nil {
 		return err
 	}
 
-	app.loki.info(fmt.Sprintf(`done in %s; added "%s"`, formatDuration(time.Now().Sub(startTime)), q))
+	app.loki.info(fmt.Sprintf(`Done in %s; added "%s"`, formatDuration(time.Now().Sub(startTime)), q))
 	return nil
 }
 
-func sendSearchBodyFromDDGToQueue(app App, proxyURL, userAgent string) error {
-	chSource, err := app.rabbitConn.Channel()
-	if err != nil {
-		return err
-	}
-	defer func(chSource *amqp.Channel) {
-		if err := chSource.Close(); err != nil {
-			app.loki.error(errors.WithStack(err))
-		}
-	}(chSource)
-
-	chDest, err := app.rabbitConn.Channel()
-	if err != nil {
-		return err
-	}
-	defer func(chDest *amqp.Channel) {
-		if err := chDest.Close(); err != nil {
-			app.loki.error(errors.WithStack(err))
-		}
-	}(chDest)
-
-	handler := func(in []byte) error {
+func sendSearchBodyFromDDGToQueue(app App, queueSource, queueDest *RabbitMQSession, proxyURL, userAgent string) {
+	var handler RabbitMQStreamHandler = func(in []byte) (ack bool) {
 		startTime := time.Now()
 
 		searchURL := makeDDGSearchURL(string(in))
@@ -86,7 +55,8 @@ func sendSearchBodyFromDDGToQueue(app App, proxyURL, userAgent string) error {
 
 		resp, err := sendRequest(searchURL, proxyURL, headers)
 		if err != nil {
-			return err
+			app.loki.debug(fmt.Sprintf(`sendRequest("%s", "%s", headers); err = %s`, searchURL, proxyURL, errors.WithStack(err)))
+			return false
 		}
 		defer func(Body io.ReadCloser) {
 			if err := Body.Close(); err != nil {
@@ -95,51 +65,29 @@ func sendSearchBodyFromDDGToQueue(app App, proxyURL, userAgent string) error {
 		}(resp.Body)
 
 		if resp.StatusCode != http.StatusOK {
-			return err
+			app.loki.debug(fmt.Sprintf(`sendRequest("%s", "%s", headers); resp.StatusCode = %d`, searchURL, proxyURL, resp.StatusCode))
+			return false
 		}
 		body, err := ioutil.ReadAll(resp.Body)
 		if err != nil {
-			return err
+			app.loki.debug(fmt.Sprintf(`sendRequest("%s", "%s", headers); can't read resp.Body`, searchURL, proxyURL))
+			return false
 		}
 
-		if err := publish(chDest, queueSearchBodiesFromDDG, body); err != nil {
+		if err := queueDest.Push(body); err != nil {
 			app.loki.error(errors.WithStack(err))
-			return err
+			return false
 		}
 
-		app.loki.info(fmt.Sprintf(`done in %s; added body from "%s"`, formatDuration(time.Now().Sub(startTime)), searchURL))
-		return nil
+		app.loki.info(fmt.Sprintf(`Done in %s; added body from "%s"`, formatDuration(time.Now().Sub(startTime)), searchURL))
+		return true
 	}
 
-	if err := consume(chSource, queueSearchQueries, handler); err != nil {
-		return err
-	}
-
-	return nil
+	queueSource.handleStream(handler)
 }
 
-func processSearchBodyFromDDG(app App) error {
-	chSource, err := app.rabbitConn.Channel()
-	if err != nil {
-		return err
-	}
-	defer func(chSource *amqp.Channel) {
-		if err := chSource.Close(); err != nil {
-			app.loki.error(errors.WithStack(err))
-		}
-	}(chSource)
-
-	chDest, err := app.rabbitConn.Channel()
-	if err != nil {
-		return err
-	}
-	defer func(chDest *amqp.Channel) {
-		if err := chDest.Close(); err != nil {
-			app.loki.error(errors.WithStack(err))
-		}
-	}(chDest)
-
-	handler := func(in []byte) error {
+func processSearchBodyFromDDG(app App, queueSource, queueDest *RabbitMQSession) {
+	var handler RabbitMQStreamHandler = func(in []byte) (ack bool) {
 		startTime := time.Now()
 
 		urls := findSiteURLsFromDDG(string(in))
@@ -148,51 +96,27 @@ func processSearchBodyFromDDG(app App) error {
 			changeType, err := app.rdbForSites.set(u)
 			if err != nil {
 				app.loki.error(errors.WithStack(err))
-				return err
+				return false
 			}
 			switch changeType {
 			case redisChangeAdd, redisChangeUpdate:
-				if err := publish(chDest, queueProxySources, []byte(u)); err != nil {
+				if err := queueDest.Push([]byte(u)); err != nil {
 					app.loki.error(errors.WithStack(err))
-					return err
+					return false
 				}
 				addedUrls += 1
 			}
 		}
 
-		app.loki.info(fmt.Sprintf(`done in %s; handled %d urls, added %d urls`, formatDuration(time.Now().Sub(startTime)), len(urls), addedUrls))
-		return nil
+		app.loki.info(fmt.Sprintf(`Done in %s; handled %d urls, added %d urls`, formatDuration(time.Now().Sub(startTime)), len(urls), addedUrls))
+		return true
 	}
 
-	if err := consume(chSource, queueSearchBodiesFromDDG, handler); err != nil {
-		return err
-	}
-
-	return nil
+	queueSource.handleStream(handler)
 }
 
-func sendHTMLFromProxySourceToQueue(app App) error {
-	chSource, err := app.rabbitConn.Channel()
-	if err != nil {
-		return err
-	}
-	defer func(chSource *amqp.Channel) {
-		if err := chSource.Close(); err != nil {
-			app.loki.error(errors.WithStack(err))
-		}
-	}(chSource)
-
-	chDest, err := app.rabbitConn.Channel()
-	if err != nil {
-		return err
-	}
-	defer func(chDest *amqp.Channel) {
-		if err := chDest.Close(); err != nil {
-			app.loki.error(errors.WithStack(err))
-		}
-	}(chDest)
-
-	handler := func(in []byte) error {
+func sendHTMLFromProxySourceToQueue(app App, queueSource, queueDest *RabbitMQSession) {
+	var handler RabbitMQStreamHandler = func(in []byte) (ack bool) {
 		startTime := time.Now()
 
 		var headers = map[string]string{
@@ -203,7 +127,8 @@ func sendHTMLFromProxySourceToQueue(app App) error {
 		u := string(in)
 		resp, err := sendRequest(u, proxyURL, headers)
 		if err != nil {
-			return err
+			app.loki.debug(fmt.Sprintf(`sendRequest("%s", "%s", headers); err = %s`, u, proxyURL, errors.WithStack(err)))
+			return false
 		}
 		defer func(Body io.ReadCloser) {
 			if err := Body.Close(); err != nil {
@@ -212,11 +137,13 @@ func sendHTMLFromProxySourceToQueue(app App) error {
 		}(resp.Body)
 
 		if resp.StatusCode != http.StatusOK {
-			return err
+			app.loki.debug(fmt.Sprintf(`sendRequest("%s", "%s", headers); resp.StatusCode = %d`, u, proxyURL, resp.StatusCode))
+			return false
 		}
 		body, err := ioutil.ReadAll(resp.Body)
 		if err != nil {
-			return err
+			app.loki.debug(fmt.Sprintf(`sendRequest("%s", "%s", headers); can't read resp.Body`, u, proxyURL))
+			return false
 		}
 		hPayload := htmlPayload{
 			HTML:    string(body),
@@ -224,67 +151,35 @@ func sendHTMLFromProxySourceToQueue(app App) error {
 		}
 		hPayloadJSON, err := json.Marshal(hPayload)
 		if err != nil {
-			return err
+			app.loki.error(fmt.Errorf(`can't json.Marshal(payload) from url "%s": %s`, u, errors.WithStack(err)))
+			return false
 		}
 
-		if err := publish(chDest, queueProxySourceHTML, hPayloadJSON); err != nil {
+		if err := queueDest.Push(hPayloadJSON); err != nil {
 			app.loki.error(errors.WithStack(err))
+			return false
 		}
 
-		app.loki.info(fmt.Sprintf(`done in %s; added body from "%s"`, formatDuration(time.Now().Sub(startTime)), u))
-		return nil
+		app.loki.info(fmt.Sprintf(`Done in %s; added body from "%s"`, formatDuration(time.Now().Sub(startTime)), u))
+		return true
 	}
 
-	if err := consume(chSource, queueProxySources, handler); err != nil {
-		return err
-	}
-
-	return nil
+	queueSource.handleStream(handler)
 }
 
-func processSourceHTML(app App) error {
-	chSource, err := app.rabbitConn.Channel()
-	if err != nil {
-		return err
-	}
-	defer func(chSource *amqp.Channel) {
-		if err := chSource.Close(); err != nil {
-			app.loki.error(errors.WithStack(err))
-		}
-	}(chSource)
-
-	chDestRawProxies, err := app.rabbitConn.Channel()
-	if err != nil {
-		return err
-	}
-	defer func(chDestRawProxies *amqp.Channel) {
-		if err := chDestRawProxies.Close(); err != nil {
-			app.loki.error(errors.WithStack(err))
-		}
-	}(chDestRawProxies)
-
-	chDestProxySources, err := app.rabbitConn.Channel()
-	if err != nil {
-		return err
-	}
-	defer func(chDestProxySources *amqp.Channel) {
-		if err := chDestProxySources.Close(); err != nil {
-			app.loki.error(errors.WithStack(err))
-		}
-	}(chDestProxySources)
-
+func processSourceHTML(app App, queueSource, queueDestRawProxies, queueDestProxySources *RabbitMQSession) {
 	forbiddenDomains, err := blacklistDomains(app.postgresPool)
 	if err != nil {
-		return err
+		app.loki.error(errors.WithStack(err))
 	}
 
-	handler := func(in []byte) error {
+	var handler RabbitMQStreamHandler = func(in []byte) (ack bool) {
 		startTime := time.Now()
 
 		var hPayload htmlPayload
 		if err := json.Unmarshal(in, &hPayload); err != nil {
-			app.loki.error(errors.WithStack(err))
-			return err
+			app.loki.error(fmt.Errorf(`can't json.Unmarshal(): %s`, errors.WithStack(err)))
+			return false
 		}
 
 		proxies := findProxiesFromHTML(hPayload.HTML)
@@ -293,13 +188,13 @@ func processSourceHTML(app App) error {
 			changeType, err := app.rdbForProxies.set(p)
 			if err != nil {
 				app.loki.error(errors.WithStack(err))
-				return err
+				return false
 			}
 			switch changeType {
 			case redisChangeAdd, redisChangeUpdate:
-				if err := publish(chDestRawProxies, queueRawProxies, []byte(p)); err != nil {
+				if err := queueDestRawProxies.Push([]byte(p)); err != nil {
 					app.loki.error(errors.WithStack(err))
-					return err
+					return false
 				}
 				addedProxies += 1
 			}
@@ -319,50 +214,36 @@ func processSourceHTML(app App) error {
 				changeType, err := app.rdbForSites.set(u)
 				if err != nil {
 					app.loki.error(errors.WithStack(err))
-					return err
+					return false
 				}
 				switch changeType {
 				case redisChangeAdd, redisChangeUpdate:
-					if err := publish(chDestProxySources, queueProxySources, []byte(u)); err != nil {
+					if err := queueDestProxySources.Push([]byte(u)); err != nil {
 						app.loki.error(errors.WithStack(err))
-						return err
+						return false
 					}
 					addedURLs += 1
 				}
 			}
 		}
 
-		app.loki.info(fmt.Sprintf(`done in %s; processed "%s"; found/added: %d/%d proxies, %d/%d urls`, formatDuration(time.Now().Sub(startTime)), hPayload.FromURL, len(proxies), addedProxies, len(urls), addedURLs))
-		return nil
+		app.loki.info(fmt.Sprintf(`Done in %s; processed "%s"; found/added: %d/%d proxies, %d/%d urls`, formatDuration(time.Now().Sub(startTime)), hPayload.FromURL, len(proxies), addedProxies, len(urls), addedURLs))
+		return true
 	}
 
-	if err := consume(chSource, queueProxySourceHTML, handler); err != nil {
-		return err
-	}
-
-	return nil
+	queueSource.handleStream(handler)
 }
 
-func processRawProxy(app App) error {
-	ch, err := app.rabbitConn.Channel()
-	if err != nil {
-		return err
-	}
-	defer func(ch *amqp.Channel) {
-		if err := ch.Close(); err != nil {
-			app.loki.error(errors.WithStack(err))
-		}
-	}(ch)
-
-	handler := func(in []byte) error {
+func processRawProxy(app App, queueSource *RabbitMQSession) {
+	var handler RabbitMQStreamHandler = func(in []byte) (ack bool) {
 		startTime := time.Now()
 
 		pURL := string(in)
 
 		pType, anonymous, err := checkProxy(pURL, app.conf.IPAPIURL)
 		if err != nil {
-			app.loki.info(fmt.Sprintf(`done in %s; discarded "%s"`, formatDuration(time.Now().Sub(startTime)), pURL))
-			return err
+			app.loki.info(fmt.Sprintf(`Done in %s; discarded "%s"`, formatDuration(time.Now().Sub(startTime)), pURL))
+			return false
 		}
 
 		p := proxy{
@@ -376,59 +257,35 @@ func processRawProxy(app App) error {
 
 		if err := saveProxyToDB(app.postgresPool, p); err != nil {
 			app.loki.error(errors.WithStack(err))
-			return err
+			return false
 		}
 
-		app.loki.info(fmt.Sprintf(`done in %s; added "%s"`, formatDuration(time.Now().Sub(startTime)), pURL))
-		return nil
+		app.loki.info(fmt.Sprintf(`Done in %s; added "%s"`, formatDuration(time.Now().Sub(startTime)), pURL))
+		return true
 	}
 
-	if err := consume(ch, queueRawProxies, handler); err != nil {
-		return err
-	}
-
-	return nil
+	queueSource.handleStream(handler)
 }
 
-func fillCheckProxiesQueue(app App) error {
-	ch, err := app.rabbitConn.Channel()
-	if err != nil {
-		return err
-	}
-	defer func(ch *amqp.Channel) {
-		if err := ch.Close(); err != nil {
-			app.loki.error(errors.WithStack(err))
-		}
-	}(ch)
-
+func fillCheckProxiesQueue(app App, queueDest *RabbitMQSession) error {
 	startTime := time.Now()
 
-	proxies, err := freshProxies(app.postgresPool, time.Hour*24)
+	proxies, err := freshProxies(app.postgresPool, time.Hour*24*30) // 30 days
 	if err != nil {
 		return err
 	}
 	for _, p := range proxies {
-		if err := publish(ch, queueCheckProxies, []byte(p.URL)); err != nil {
+		if err := queueDest.Push([]byte(p.URL)); err != nil {
 			return err
 		}
 	}
 
-	app.loki.info(fmt.Sprintf(`done in %s; added %d proxies`, formatDuration(time.Now().Sub(startTime)), len(proxies)))
+	app.loki.info(fmt.Sprintf(`Done in %s; added %d proxies`, formatDuration(time.Now().Sub(startTime)), len(proxies)))
 	return nil
 }
 
-func processCheckProxy(app App) error {
-	ch, err := app.rabbitConn.Channel()
-	if err != nil {
-		return err
-	}
-	defer func(ch *amqp.Channel) {
-		if err := ch.Close(); err != nil {
-			app.loki.error(errors.WithStack(err))
-		}
-	}(ch)
-
-	handler := func(in []byte) error {
+func processCheckProxy(app App, queueSource *RabbitMQSession) {
+	var handler RabbitMQStreamHandler = func(in []byte) (ack bool) {
 		startTime := time.Now()
 
 		pURL := string(in)
@@ -437,10 +294,10 @@ func processCheckProxy(app App) error {
 		if err != nil {
 			if err := changeProxyToInactive(app.postgresPool, pURL); err != nil {
 				app.loki.error(errors.WithStack(err))
-				return err
+				return false
 			}
-			app.loki.info(fmt.Sprintf(`done in %s; changed to inactive "%s"`, formatDuration(time.Now().Sub(startTime)), pURL))
-			return err
+			app.loki.info(fmt.Sprintf(`Done in %s; changed to inactive "%s"`, formatDuration(time.Now().Sub(startTime)), pURL))
+			return false
 		}
 
 		p := proxy{
@@ -453,16 +310,12 @@ func processCheckProxy(app App) error {
 
 		if err := updateProxyInDB(app.postgresPool, p); err != nil {
 			app.loki.error(errors.WithStack(err))
-			return err
+			return false
 		}
 
-		app.loki.info(fmt.Sprintf(`done in %s; updated last check (active) "%s"`, formatDuration(time.Now().Sub(startTime)), pURL))
-		return nil
+		app.loki.info(fmt.Sprintf(`Done in %s; updated last check (active) "%s"`, formatDuration(time.Now().Sub(startTime)), pURL))
+		return true
 	}
 
-	if err := consume(ch, queueCheckProxies, handler); err != nil {
-		return err
-	}
-
-	return nil
+	queueSource.handleStream(handler)
 }